ci tests are green!!

+++ bin/run_tests.sh
@@ -21,5 +21,7 @@ fi
 # Run integration tests
 node node_modules/cucumber/bin/cucumber.js features
 
-# Do some cleanup
-kill $(jobs -p)
+# Do some cleanup if running locally
+if [ -z "$CI" ]; then
+  kill $(jobs -p)
+fi
